Validate found path
Both A* and JPS allow a diagonal movement between adjacent nodes when
one of the horizontal or vertical cell in the diagonal is blocked.
However, that movement is not really valid beacause an entity will
collide with the blocked cell. Therefore, the validation just returns a
valid node for a pair of consecutive nodes in the path